Fix monitor comparisons in AVL trees
The monitor comparison logic is broken due to the possibility that the
64 bit subtraction can leave a positive value that becomes negative when
down casted to a signed 32 bit integer. The end result is that a
process's monitor AVL tree enters an invalid state which then prevents
monitor signals from being delivered.

This fixes the issue by returning hard coded numerical values rather
than the difference that is down casted.