TMP - Hacks to reproduce broken monitor delivery
There are three main bits that I've added here. First is to extend
`erts_debug:set_internal_state/2` to allow for setting the `ref` field
in `ErlSchedulerData` structs. Second, I've added a validation function
that walks the monitor AVL tree and verifies that its ordered
correctly. And lastly are a couple functions to limit the assertions to
a specific process (namely one that has done a `put(registered_name,
badmons)`.