[erts] Fix fetch of signals in parallel enqueue race scenario
When fetching signals for a process not in a running state and with parallel
signal enqueue enabled, it could end up being activated but not inserted into
the run queue. If/when this happened the process would be stuck potentially
for ever.

Currently signals are only fetched for processes not in a running state
when a process is receive traced.